Support different threshold fractions for lo/hi
The `-t` option specifies the fraction of samples to "clip" at the low
and high extremes of the input sample range.  This option may now be
given multiple times with the first use specifying the fraction of low
samples to clip and the last use specifying the fraction of high samples
to clip.  As before, if used one time, the fraction given is used for
both ends and if not given then 3f-6 (0.000006) is used for both ends.