Add bind address to host specs in setup_workers
This ensures that master-worker communication happens over the same
network interface as the SSH connection, which is useful when the SSH
connection is over a high speed network (i.e. >1 GbE).  Without the
explicit bind address in the host spec, Distributed.jl will use the
worker's first interface (often 1 GbE) for master-worker communication.